Created a function that searches for available space within a given range + begun playing around with spawning furniture.

+++ Assets/Jacob/Tile.cs
@@ -16,8 +16,8 @@ public class Tile{
 	public float damage;//The damage value of the tile. If positive, the tile will inflict damage on the player. If negative, the tile will heal the player.
 	public float speed;//If 1, no change on player speed. If less than one, player speed will be reduced. If bigger than one, player speed will be increased. (Through multiplication)
 
-	public float x;//This will hold the x position of the tile
-	public float y;//This will hold the y position of the tile
+	public int x;//This will hold the x position of the tile
+	public int y;//This will hold the y position of the tile
 
 	/// <summary>
 	/// Initializes a new instance of the <see cref="Tile"/> class.
@@ -32,7 +32,7 @@ public class Tile{
 		y = 0;
 	}
 
-	public Tile(float x, float y){//Default constructor
+	public Tile(int x, int y){//Default constructor
 		canSpawnEnemies = false;
 		isWalkable = false;
 		damage = 0;
@@ -40,7 +40,7 @@ public class Tile{
 		this.x = x;
 		this.y = y;
 	}
-	public Tile(float x, float y, float speed, float damage, bool isWalkable, bool canSpawnEnemies){//more specified constructor
+	public Tile(int x, int y, float speed, float damage, bool isWalkable, bool canSpawnEnemies){//more specified constructor
 		this.canSpawnEnemies = canSpawnEnemies;
 		this.speed = speed;
 		this.damage = damage;
